# docs/smart_court_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SmartCourt API",
    "version": "1.0.0",
    "description": "SmartCourt Frontend API Reference. This document is traced from the controller, request/response DTO, validator, and service/handler implementations currently under SmartCourt/Features. It describes the 25 routes exposed by those controllers. All validation messages, field constraints, Arabic error descriptions, and rate limit rules match docs/frontend_api.md exactly."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Send Authorization: Bearer <access-token> for secured routes."
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indicates whether the request succeeded."
          },
          "message": {
            "type": "string",
            "nullable": true,
            "description": "Human-readable message or error description."
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true,
            "description": "List of specific error messages if any."
          },
          "statusCode": {
            "type": "integer",
            "description": "HTTP status code."
          }
        },
        "required": [
          "success",
          "statusCode"
        ]
      },
...
      "VerificationDocumentDetailsDto": {
        "type": "object",
        "properties": {
          "documentId": {
            "type": "string",
            "format": "uuid"
          },
          "documentType": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "expirationDate": {
            "type": "string",
            "format": "date"
          },
          "reviewedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "rejectionReason": {
            "type": "string",
            "nullable": true
          },
          "contentUrl": {
            "type": "string"
          }
        },
        "required": [
          "documentId",
          "documentType",
          "status",
          "fileName",
          "contentType",
          "expirationDate",
          "contentUrl"
        ]
      },
      "VerificationDetailsDto": {
        "type": "object",
        "properties": {
          "lawyerId": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "accountStatus": {
            "type": "string"
          },
          "isFullyVerified": {
            "type": "boolean"
          },
          "documents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VerificationDocumentDetailsDto"
            }
          }
        },
        "required": [
          "lawyerId",
          "fullName",
          "email",
          "accountStatus",
          "isFullyVerified",
          "documents"
        ]
      },
...
      "UploadedDocumentDto": {
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string"
          },
          "type": {
            "type": "integer",
            "description": "1=NationalIdFront, 2=NationalIdBack, 3=BarAssociationCardFront, 4=BarAssociationCardBack"
          }
        },
        "required": [
          "fileName",
          "type"
        ]
      },
      "DocumentUploadErrorDto": {
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string"
          },
          "type": {
            "type": "integer"
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "fileName",
          "type",
          "error"
        ]
      },
      "SubmitVerificationDocumentResponseDto": {
        "type": "object",
        "properties": {
          "uploadedDocuments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UploadedDocumentDto"
            }
          },
          "failedDocuments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentUploadErrorDto"
            }
          }
        },
        "required": [
          "uploadedDocuments",
          "failedDocuments"
        ]
      },
      "UserVerificationDocumentDto": {
        "type": "object",
        "properties": {
          "documentId": {
            "type": "string",
            "format": "uuid"
          },
          "documentType": {
            "type": "integer",
            "description": "1=NationalIdFront, 2=NationalIdBack, 3=BarAssociationCardFront, 4=BarAssociationCardBack"
          },
          "status": {
            "type": "integer",
            "description": "1=Pending, 2=Verified, 3=Rejected, 4=Expired"
          },
          "expirationDate": {
            "type": "string",
            "format": "date"
          },
          "isCurrent": {
            "type": "boolean"
          },
          "fileName": {
            "type": "string"
          }
        },
        "required": [
          "documentId",
          "documentType",
          "status",
          "expirationDate",
          "isCurrent",
          "fileName"
        ]
      },
      "GetUserVerificationDocumentsResponseDto": {
        "type": "object",
        "properties": {
          "documents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserVerificationDocumentDto"
            }
          }
        },
        "required": [
          "documents"
        ]
      },
...
    }
  },
...
  "paths": {
...
    "/api/auth/confirm-email": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Confirm email",
        "description": "Confirms an email address from the user id and Base64URL-encoded confirmation token. Rate limit: IP 20/15 minutes and account key 5/hour.",
        "security": [],
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 64
            },
            "description": "User ID Guid"
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 2048
            },
            "description": "Base64URL-encoded token"
          }
        ],
        "responses": {
          "200": {
            "description": "Email confirmed successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "example": {
                  "success": true,
                  "message": "تم تأكيد البريد الإلكتروني بنجاح.",
                  "errors": null,
                  "statusCode": 200
                }
              }
            }
          },
          "400": {
            "description": "BusinessException('رابط تأكيد البريد الإلكتروني غير صالح أو منتهي الصلاحية.')"
          },
          "429": {
            "description": "Rate limit exceeded."
          }
        }
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Forgot password",
        "description": "Requests a password-reset email. Rate limit: IP 5/15 minutes and account key 3/hour.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Required; valid email (عنوان البريد الإلكتروني مطلوب / عنوان البريد الإلكتروني غير صالح)."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset email sent if eligible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "example": {
                  "success": true,
                  "message": "إذا كان البريد الإلكتروني مسجلاً، سيتم إرسال رابط إعادة تعيين كلمة المرور",
                  "errors": null,
                  "statusCode": 200
                }
              }
            }
          },
          "400": {
            "description": "Validation failure."
          },
          "429": {
            "description": "Rate limit exceeded."
          },
          "500": {
            "description": "Queue email failure."
          }
        }
      }
    },
...
    "/api/admin/verifications/{lawyerId}": {
      "get": {
        "tags": [
          "Admin verifications"
        ],
        "summary": "Get lawyer verification details",
        "description": "Returns the lawyer's current verification documents and account verification state. Requires Admin role.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "lawyerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lawyer verification details.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/VerificationDetailsDto"
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "success": true,
                  "data": {
                    "lawyerId": "0f2c5a4d-7a3e-4b7a-8d13-2ef7df7d54f2",
                    "fullName": "Example Lawyer",
                    "email": "lawyer@example.com",
                    "phoneNumber": "+201012345678",
                    "accountStatus": "PendingReview",
                    "isFullyVerified": false,
                    "documents": [
                      {
                        "documentId": "7ed88c0a-7e67-4a34-bdfd-7b0a6dfb4e3a",
                        "documentType": "NationalIdFront",
                        "status": "Pending",
                        "fileName": "national-id-front.jpg",
                        "contentType": "image/jpeg",
                        "expirationDate": "2030-12-31",
                        "reviewedAt": null,
                        "rejectionReason": null,
                        "contentUrl": "/api/admin/verifications/documents/7ed88c0a-7e67-4a34-bdfd-7b0a6dfb4e3a/content"
                      }
                    ]
                  },
                  "message": null,
                  "errors": null,
                  "statusCode": 200
                }
              }
            }
          },
          "400": {
            "description": "Empty GUID validation error."
          },
          "404": {
            "description": "NotFoundException('Lawyer was not found.')"
          }
        }
      }
    },
...
    "/api/UserVerification/submit-verification-documents": {
      "post": {
        "tags": [
          "User verification documents"
        ],
        "summary": "Submit verification documents",
        "description": "Uploads one or more verification images for a user and returns per-file successes and failures. Anonymous controller action.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "userId",
                  "documents[0].file",
                  "documents[0].expirationDate",
                  "documents[0].type"
                ],
                "properties": {
                  "userId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "Required (UserId is required)"
                  },
                  "documents[0].file": {
                    "type": "string",
                    "format": "binary",
                    "description": "JPEG, PNG, WEBP, HEIC, or HEIF image"
                  },
                  "documents[0].expirationDate": {
                    "type": "string",
                    "format": "date",
                    "description": "Must be future date"
                  },
                  "documents[0].type": {
                    "type": "integer",
                    "enum": [
                      1,
                      2,
                      3,
                      4,
                      5
                    ],
                    "description": "1=NationalIdFront, 2=NationalIdBack, 3=BarAssociationCardFront, 4=BarAssociationCardBack, 5=other"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Documents uploaded.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/SubmitVerificationDocumentResponseDto"
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "success": true,
                  "data": {
                    "uploadedDocuments": [
                      {
                        "fileName": "id-front.jpg",
                        "type": 1
                      }
                    ],
                    "failedDocuments": [
                      {
                        "fileName": "old-card.jpg",
                        "type": 3,
                        "error": "This document is expired"
                      }
                    ]
                  },
                  "message": null,
                  "errors": null,
                  "statusCode": 200
                }
              }
            }
          },
          "400": {
            "description": "Validation or upload error."
          }
        }
      }
    },
    "/api/UserVerification/{UserId}": {
      "get": {
        "tags": [
          "User verification documents"
        ],
        "summary": "Get user verification documents",
        "description": "Lists all verification documents belonging to the supplied user id. Anonymous controller action.",
        "security": [],
        "parameters": [
          {
            "name": "UserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User verification documents returned.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/GetUserVerificationDocumentsResponseDto"
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "success": true,
                  "data": {
                    "documents": [
                      {
                        "documentId": "7ed88c0a-7e67-4a34-bdfd-7b0a6dfb4e3a",
                        "documentType": 1,
                        "status": 1,
                        "expirationDate": "2030-12-31",
                        "isCurrent": true,
                        "fileName": "id-front.jpg"
                      }
                    ]
                  },
                  "message": null,
                  "errors": null,
                  "statusCode": 200
                }
              }
            }
          },
          "400": {
            "description": "User Id is required."
          },
          "404": {
            "description": "The specified user does not exist."
          }
        }
      }
    },
    "/api/UserVerification": {
      "delete": {
        "tags": [
          "User verification documents"
        ],
        "summary": "Delete verification document",
        "description": "Deletes a user's verification document from storage and the database. Anonymous controller action.",
        "security": [],
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "documentId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "example": {
                  "success": true,
                  "message": null,
                  "errors": null,
                  "statusCode": 200
                }
              }
            }
          },
          "400": {
            "description": "Storage deletion failure or validation error."
          },
          "404": {
            "description": "User or verification document not found."
          }
        }
      }
    },
...
  }
}